tec: Rename registration-form controller in csrf-loader
