+++ src/assets/javascripts/controllers/csrf_loader_controller.js
